+++ openmc_dagmc_wrapper/tallies/__init__.py
@@ -1,1 +1,2 @@
 from .cell_tally import CellTally, CellTallies
+from .tet_mesh_tallies import TetMeshTally, TetMeshTallies
